requirements.txt
beautifulsoup4==latest
python-dotenv==latest
requests==latest
selenium==latest